+++ RealEstateMVC_NOAUTH/Controllers/Admin/ContactsController.cs
@@ -13,16 +13,29 @@ namespace RealEstateMVC_NOAUTH.Controllers.Admin
     public class ContactsController : Controller
     {
         private RealEstateEntities db = new RealEstateEntities();
+        private bool IsLogin()
+        {
+            return (Session.Count > 0 && Session["Type"].Equals("Admin"));
+        }
 
         // GET: Contacts
         public ActionResult Index()
         {
+            if (!IsLogin())
+            {
+                return RedirectToAction("Index", "Admin");
+            }
             return View(db.CONTACTs.ToList());
         }
 
         // GET: Contacts/Delete/5
         public ActionResult Delete(int? id)
         {
+            if (!IsLogin())
+            {
+                return RedirectToAction("Index", "Admin");
+            }
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -40,6 +53,11 @@ namespace RealEstateMVC_NOAUTH.Controllers.Admin
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (!IsLogin())
+            {
+                return RedirectToAction("Index", "Admin");
+            }
+
             CONTACT cONTACT = db.CONTACTs.Find(id);
             db.CONTACTs.Remove(cONTACT);
             db.SaveChanges();
